Safari-611.1.10/LayoutTests/css-custom-properties-api/cycles-expected.txt

test

test

test

test

test


PASS Registration is successful
PASS JS Attributes are valid for element 1
PASS JS Attributes are valid for element 2
PASS JS Attributes are valid for element 3
PASS JS Attributes are valid for element 4
PASS JS Attributes are valid for element 5
